First pass at enabling exportable memory pools
This commit adds support to vmaCreatePool(), via its flags, to mark the
pool as exportable. All memory allocated for this pool with have the
exportable property enabled. It is the user's responsibility to ensure
that the pool is being created with the right memory type, and that
buffers and other objects bound to this memory have the appopriate
exportable properties set.